# docs/api-spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Nyamby API",
    "version": "1.0.0",
    "description": "API spec untuk MVP Nyamby — Next.js Route Handlers. Base URL: /api",
    "contact": {
      "name": "Satria Romanda"
    }
  },
  "servers": [
    { "url": "http://localhost:3000/api", "description": "Local Development" },
    { "url": "https://nyamby.vercel.app/api", "description": "Production (Vercel)" }
  ],
...
  "paths": {
...
    "/jobs": {
      "post": {
        "tags": ["Jobs"],
        "summary": "Client buat job baru",
        "description": "Menyimpan job posting dan **otomatis men-trigger AI Job Matching** terhadap semua talenta terdaftar.",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["title", "description", "category", "required_skills"],
                "properties": {
                  "title": { "type": "string", "example": "Landing Page untuk UMKM Kuliner" },
                  "description": { "type": "string", "example": "Butuh landing page responsive dengan CTA dan form kontak..." },
                  "category": { "type": "string", "enum": ["web_dev", "graphic_designer"] },
                  "budget_min": { "type": "number", "example": 1000000 },
                  "budget_max": { "type": "number", "example": 3000000 },
                  "deadline": { "type": "string", "format": "date", "example": "2026-06-30" },
                  "required_skills": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "skill_id": { "type": "string", "format": "uuid" },
                        "is_mandatory": { "type": "boolean" }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "example": {
                  "success": true,
                  "message": "Job berhasil diposting. AI sedang mencarikan talenta terbaik.",
                  "data": { "job_id": "uuid", "status": "active", "matching_triggered": true }
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": ["Jobs"],
        "summary": "List semua job aktif",
        "parameters": [
          { "name": "category", "in": "query", "schema": { "type": "string", "enum": ["web_dev", "graphic_designer"] } },
          { "name": "skill", "in": "query", "schema": { "type": "string" }, "description": "Filter by skill name" },
          { "name": "budget_max", "in": "query", "schema": { "type": "number" } },
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 10 } }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "success": true,
                  "data": {
                    "jobs": [
                      { "id": "uuid", "title": "Landing Page UMKM", "category": "web_dev", "budget_max": 3000000, "deadline": "2026-06-30", "status": "active", "required_skills": ["React", "Tailwind CSS"] }
                    ],
                    "pagination": { "page": 1, "limit": 10, "total": 45 }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/jobs/{id}/status": {
      "patch": {
        "tags": ["Jobs"],
        "summary": "Update status job",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "status": { "type": "string", "enum": ["active", "matched", "in_progress", "completed", "cancelled"] }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Status berhasil diupdate" }
        }
      }
    },
...
    "/matches/{id}/status": {
      "patch": {
        "tags": ["AI"],
        "summary": "Talenta accept atau reject job offer",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "description": "job_match ID", "schema": { "type": "string", "format": "uuid" } }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "status": { "type": "string", "enum": ["applied", "rejected"] }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Status match diupdate" }
        }
      }
    },
...
    "/notifications": {
      "get": {
        "tags": ["Notifications"],
        "summary": "Ambil notifikasi user yang sedang login",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "unread_only", "in": "query", "schema": { "type": "boolean", "default": false } }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "success": true,
                  "data": {
                    "unread_count": 2,
                    "notifications": [
                      { "id": "uuid", "type": "new_match", "message": "Ada job baru yang cocok untukmu: Landing Page UMKM (87% match)", "is_read": false, "created_at": "2026-05-28T10:00:00Z" }
                    ]
                  }
                }
              }
            }
          }
        }
      }
    },
    "/notifications/{id}/read": {
      "patch": {
        "tags": ["Notifications"],
        "summary": "Mark notifikasi sebagai dibaca",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": { "description": "Notifikasi ditandai sebagai dibaca" }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token dari NextAuth.js. Sertakan di header: Authorization: Bearer <token>"
      }
    }
  }
}